upa-portal/backend_backup/app/pyproject.toml
[tool.poetry]
name = "adjuster-portal"
version = "0.1.0"
description = "Adjuter Portal for the Adjuster Agents."
authors = ["Navi Singh <navisingh@live.in>"]
readme = "README.md"
packages = [{ include = "adjuster_portal" }]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.22.0"
tenacity = "^8.2.2"
sqlalchemy = { extras = ["mypy"], version = "^2.0.13" }
pydantic = { extras = ["email"], version = "^1.10.7" }
psycopg2-binary = "^2.9.6"
celery = "^5.2.7"
fastapi = "^0.95.1"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
alembic = "^1.10.4"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
python-multipart = "^0.0.6"
jinja2 = "^3.1.2"
boto3 = "^1.26.163"
pillow = "^9.5.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.2.0"
black = "^23.3.0"
isort = "^5.12.0"
autoflake = "^2.1.1"
flake8 = "^6.0.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.3.1"
coverage = "^7.2.5"
httpx = "^0.24.0"

[tool.isort]
profile = "black"

[tool.autoflake]
exclude = "__init__.py"
recursive = true
remove-all-unused-imports = true
remove-unused-variables = true
in-place = true

[tool.mypy]
plugins = "pydantic.mypy, sqlalchemy.ext.mypy.plugin"
ignore_missing_imports = true
disallow_untyped_defs = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
